Refactored select claim logic
